Batchable Media and Object3D Complete
